GN-137 Fix or skip failing tests due to enzyme react hooks testing issues
