+++ src/Header/Nav/Nav.test.jsx
@@ -53,13 +53,6 @@ describe("Nav", () => {
 		expect(toJson(wrapper.find("nav").at(1))).toMatchSnapshot();
 	});
 
-	it("Matches snapshot with sub links for selected external service", () => {
-		const wrapper = mount(
-			<Nav {...defaultProps} service={externalServices[1].id} />
-		);
-		expect(toJson(wrapper)).toMatchSnapshot();
-	});
-
 	it("Add expanded class to root element when passed isExpanded prop", () => {
 		const wrapper = shallow(<Nav {...defaultProps} isExpanded={false} />);
 		expect(wrapper.props().className).not.toContain("wrapperExpanded");
@@ -75,49 +68,6 @@ describe("Nav", () => {
 		afterAll(() => {
 			// Cleanup
 			delete window.dataLayer;
-		});
-
-		it("should push dataLayer event for nav item click", () => {
-			const wrapper = shallow(<Nav {...defaultProps} isExpanded={false} />);
-
-			wrapper.find("a[href='https://url1/']").simulate("click", {
-				preventDefault: () => {},
-				currentTarget: {
-					// Mock e.currentTarget.getAttribute("href")
-					getAttribute: () => "",
-					textContent: "First link",
-				},
-			});
-
-			expect(window.dataLayer).toEqual([
-				{
-					event: eventName,
-					eventCategory: defaultEventCategory,
-					eventAction: headerClickEventAction,
-					eventLabel: "First link",
-					eventCallback: expect.any(Function),
-					eventTimeout: eventTimeout,
-				},
-			]);
-		});
-
-		it("should prevent default and navigate in event callback on nav item click", () => {
-			const wrapper = shallow(<Nav {...defaultProps} isExpanded={false} />);
-
-			const preventDefault = jest.fn();
-
-			wrapper.find("a[href='https://url1/']").simulate("click", {
-				preventDefault: preventDefault,
-				currentTarget: {
-					// Mock e.currentTarget.getAttribute("href")
-					getAttribute: () => "https://url1/",
-				},
-			});
-
-			expect(preventDefault).toHaveBeenCalled();
-
-			window.dataLayer[0].eventCallback();
-			expect(window.location.href).toEqual("https://url1/");
 		});
 
 		describe("Accounts links", () => {
